[enh] Move view edge services to new service architecture
