+++ backend/services/syson-diagram-services/src/main/java/org/eclipse/syson/diagram/services/DiagramQueryGraphicalService.java
@@ -13,6 +13,7 @@
 package org.eclipse.syson.diagram.services;
 
 import java.util.List;
+import java.util.Objects;
 
 import org.eclipse.sirius.components.diagrams.elements.NodeElementProps;
 import org.eclipse.sirius.components.diagrams.renderer.DiagramRenderingCache;
@@ -47,6 +48,29 @@ public class DiagramQueryGraphicalService {
     }
 
     /**
+     * Returns {@code true} if both elements are contained in the same graphical container.
+     *
+     * @param source
+     *            the graphical source element
+     * @param target
+     *            the graphical target element
+     * @param cache
+     *            the rendering cache used in the current rendering process
+     * @return {@code true} if both elements share the same graphical container
+     */
+    public boolean isInSameGraphicalContainer(org.eclipse.sirius.components.representations.Element source,
+            org.eclipse.sirius.components.representations.Element target, DiagramRenderingCache cache) {
+        if (source.getProps() instanceof NodeElementProps sourceNodeProps
+                && target.getProps() instanceof NodeElementProps targetNodeProps) {
+            String sourceParentId = this.getParentId(cache, sourceNodeProps);
+            String targetParentId = this.getParentId(cache, targetNodeProps);
+            return (sourceParentId == null && targetParentId == null)
+                    || (sourceParentId != null && sourceParentId.equals(targetParentId));
+        }
+        return true;
+    }
+
+    /**
      * Returns {@code true} if {@code parentNodeElement} is an ancestor of {@code childNodeElement}.
      *
      * @param parentNodeElement
@@ -73,5 +97,18 @@ public class DiagramQueryGraphicalService {
             this.logger.warn("Cannot check graphical containment between {} and {}", parentNodeElement, childNodeElement);
         }
         return result;
+    }
+
+    private String getParentId(DiagramRenderingCache cache, NodeElementProps nodeElementProps) {
+        return cache.getParent(nodeElementProps.getId())
+                .map(org.eclipse.sirius.components.representations.Element::getProps)
+                .map(props -> {
+                    if (props instanceof NodeElementProps parentNodeProps) {
+                        return parentNodeProps.getId();
+                    }
+                    return null;
+                })
+                .filter(Objects::nonNull)
+                .orElse(null);
     }
 }
