+++ backend/views/syson-standard-diagrams-view/src/main/java/org/eclipse/syson/standard/diagrams/view/SDVJavaServiceProvider.java
@@ -17,7 +17,6 @@ import java.util.List;
 import org.eclipse.sirius.components.view.View;
 import org.eclipse.sirius.components.view.emf.IJavaServiceProvider;
 import org.eclipse.syson.diagram.common.view.services.ViewCreateService;
-import org.eclipse.syson.diagram.common.view.services.ViewEdgeService;
 import org.eclipse.syson.diagram.common.view.services.ViewLabelService;
 import org.eclipse.syson.diagram.services.aql.DiagramMutationAQLService;
 import org.eclipse.syson.diagram.services.aql.DiagramQueryAQLService;
@@ -47,7 +46,6 @@ public class SDVJavaServiceProvider implements IJavaServiceProvider {
         if (optGVDescription.isPresent()) {
             return List.of(DeleteService.class,
                     ViewCreateService.class,
-                    ViewEdgeService.class,
                     ViewLabelService.class,
                     UtilService.class,
                     DiagramMutationAQLService.class,
