+++ scripts/check-coverage.jsh
@@ -34,11 +34,11 @@ var moduleCoverageData = List.of(
   new ModuleCoverage("syson-application", 37.0),
   new ModuleCoverage("syson-application-configuration", 77.0),
   new ModuleCoverage("syson-common-view", 100.0),
-  new ModuleCoverage("syson-diagram-common-view", 93.0),
+  new ModuleCoverage("syson-diagram-common-view", 94.0),
   new ModuleCoverage("syson-diagram-services", 81.0),
   new ModuleCoverage("syson-direct-edit-grammar", 66.0),
   new ModuleCoverage("syson-form-services", 100.0),
-  new ModuleCoverage("syson-model-services", 95.0),
+  new ModuleCoverage("syson-model-services", 89.0),
   new ModuleCoverage("syson-representation-services", 100.0),
   new ModuleCoverage("syson-services", 70.0),
   new ModuleCoverage("syson-siriusweb-customnodes-metamodel", 41.0),
